requirements.txt
fastapi>=0.109.0
aiofiles>=23.2.1
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
google-cloud-storage>=2.14.0
httpx>=0.24.0
requests>=2.31.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
python-jose[cryptography]>=3.3.0
